+++ src/Controller/AuthController.php
@@ -5,7 +5,7 @@ namespace App\Controller;
 use App\DTO\Model\PbsUserDTO;
 use App\Model\LogMessage\SimpleLogMessage;
 use App\Service\Gamification\LoginService;
-use App\Service\DigioLogger\GelfLogger;
+use App\Service\Logger\GelfLogger;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
