+++ src/Controller/DebugController.php
@@ -4,8 +4,8 @@ namespace App\Controller;
 
 use App\Service\Aggregator\DemographicCampAggregator;
 use App\Service\Aggregator\DemographicGroupAggregator;
-use App\Service\DigioLogger\GelfLogger;
-use App\Service\DigioLogger\Messages\ExceptionLogMessage;
+use App\Service\Logger\GelfLogger;
+use App\Service\Logger\Messages\ExceptionLogMessage;
 use Exception;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
