+++ src/EventListener/ConsoleEventListener.php
@@ -5,9 +5,9 @@ namespace App\EventListener;
 use App\Command\StatisticsCommand;
 use App\Model\CommandStatistics;
 use App\Model\LogMessage\StatisticsCommandMessage;
-use App\Service\DigioLogger\GelfLogger;
-use App\Service\DigioLogger\Messages\CommandStartLogMessage;
-use App\Service\DigioLogger\Messages\ExceptionLogMessage;
+use App\Service\Logger\GelfLogger;
+use App\Service\Logger\Messages\CommandStartLogMessage;
+use App\Service\Logger\Messages\ExceptionLogMessage;
 use Symfony\Component\Console\Event\ConsoleCommandEvent;
 use Symfony\Component\Console\Event\ConsoleErrorEvent;
 use Symfony\Component\Console\Event\ConsoleTerminateEvent;
