+++ src/Model/LogMessage/StatisticsCommandMessage.php
@@ -3,7 +3,7 @@
 namespace App\Model\LogMessage;
 
 use App\Model\CommandStatistics;
-use App\Service\DigioLogger\Messages\LogMessage;
+use App\Service\Logger\Messages\LogMessage;
 
 class StatisticsCommandMessage extends LogMessage
 {
