+++ src/Service/Logger/Messages/ExceptionLogMessage.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service\DigioLogger\Messages;
+namespace App\Service\Logger\Messages;
 
 use Doctrine\ORM\EntityManagerInterface;
