+++ src/Service/Logger/Messages/RequestLogMessage.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Service\DigioLogger\Messages;
+namespace App\Service\Logger\Messages;
 
-use App\Service\DigioLogger\SecretRedactor;
+use App\Service\Logger\SecretRedactor;
 
 class RequestLogMessage extends HttpLogMessage
 {
